refactoring offer and comments controller
